Reduce precision to uint128 for currency amounts (#557)
* changed uint256 to uint128

* Added test-case for ETH App

* Reverted DOTApp and Fee source modifications

* Fixed failed test-case issue

+++ ethereum/contracts/IncentivizedOutboundChannel.sol
@@ -17,19 +17,19 @@ contract IncentivizedOutboundChannel is OutboundChannel, ChannelAccess, AccessCo
     // Nonce for last submitted message
     uint64 public nonce;
 
-    uint256 public fee;
+    uint128 public fee;
     FeeSource public feeSource;
 
     event Message(
         address source,
         uint64  nonce,
-        uint256 fee,
+        uint128 fee,
         bytes   payload
     );
 
     event FeeChanged(
-        uint256 oldFee,
-        uint256 newFee
+        uint128 oldFee,
+        uint128 newFee
     );
 
     constructor() {
@@ -55,7 +55,7 @@ contract IncentivizedOutboundChannel is OutboundChannel, ChannelAccess, AccessCo
     }
 
     // Update message submission fee.
-    function setFee(uint256 _amount) external onlyRole(CONFIG_UPDATE_ROLE) {
+    function setFee(uint128 _amount) external onlyRole(CONFIG_UPDATE_ROLE) {
         emit FeeChanged(fee, _amount);
         fee = _amount;
     }
